# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: phpcp Control Panel API
  version: "2.0.0"
  description: |
    API ของ Control Panel สำหรับบริหาร PHP Web Hosting

    **การยืนยันตัวตน** ใช้คุกกี้ session (HttpOnly + SameSite=Strict) ที่เซิร์ฟเวอร์ตั้งให้
    ไม่ใช่ bearer token — จงใจไม่ย้าย token ไปเก็บใน JavaScript เพราะจะเป็นการถอยหลัง
    จากสิ่งที่ระบบมีอยู่แล้ว (PLAN-V2 ข้อ N6)

    **CSRF** ทุกคำขอที่เปลี่ยนข้อมูล (POST/PUT/PATCH/DELETE) ต้องส่ง header
    `X-CSRF-Token` ที่ได้จาก `GET /session` · เซิร์ฟเวอร์ส่ง token ตัวใหม่กลับมาทาง
    response header ชื่อเดียวกันทุกครั้งที่มันเปลี่ยน (ตอนล็อกอิน ยืนยัน 2FA
    หรือตอน session id ถูกหมุนตามรอบ) ฝั่ง client ต้องเก็บค่าล่าสุดไว้ใช้เสมอ
servers:
  - url: https://localhost:8443/api/v2
    description: panel ของเครื่องนี้ (พอร์ตเริ่มต้น 8443)
...
paths:
...
  /backups:
    get:
      tags: [backups]
      summary: รายการไฟล์สำรอง
      description: |
        ไฟล์สำรองระดับระบบ (type `config`) ไม่ผูกกับเว็บไซต์ใด ผู้ดูแลเว็บไซต์จึงไม่เห็น
        — ข้างในมีค่า config ของทั้งเครื่อง
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PerPage" }
        - name: type
          in: query
          schema: { type: string, enum: [site, database, config, full] }
        - name: site_id
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: รายการไฟล์สำรอง
          headers:
            X-Total-Size:
              schema: { type: integer }
              description: ขนาดรวมเป็นไบต์
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, const: true }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Backup" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "401": { $ref: "#/components/responses/Error" }

    post:
      tags: [backups]
      summary: สร้างไฟล์สำรอง
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type: { type: string, enum: [site, database, config, full] }
                site_id: { type: integer, description: "ต้องระบุเมื่อ type = site" }
                database: { type: string, description: "ต้องระบุเมื่อ type = database" }
                note: { type: string, maxLength: 120 }
      responses:
        "201":
          description: สร้างแล้ว — ไฟล์พร้อมใช้ตั้งแต่ตอนที่คำตอบกลับไปถึง
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, const: true }
                  data: { $ref: "#/components/schemas/Backup" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
        "500": { $ref: "#/components/responses/Error" }
  /backups/{id}:
    delete:
      tags: [backups]
      summary: ลบไฟล์สำรอง
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: สำเร็จ — คำตอบของคำสั่งพร้อม `actions` ให้หน้าจอทำตาม
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommandResponse" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      description: ค่าที่เกินเพดานถูกบีบลงมาที่ 200 ไม่ใช่ปฏิเสธคำขอ
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: phpcp_sid
      description: |
        ตั้งโดยเซิร์ฟเวอร์เท่านั้น · HttpOnly + SameSite=Strict + Secure (เมื่อใช้ HTTPS)
        ชื่อจะเป็น `__Host-phpcp_sid` เมื่อเปิด cookie_secure
  schemas:
...
    PageMeta:
      type: object
      required: [page, per_page, total, total_pages]
      properties:
        page: { type: integer }
        per_page: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
...
    CommandResponse:
      type: object
      description: |
        คำตอบของ **คำสั่ง** (POST/PUT/PATCH/DELETE ที่เปลี่ยนสถานะ)

        **ไม่มีคีย์ `data`** — ค่าที่ผู้เรียกต้องใช้อยู่ระดับบนสุดคู่กับ `message`
        และ `actions` · กฎนี้มาจากวิธีที่ Now.js แกะคำตอบ (`response.data.data ??
        response.data`) ซึ่งจะเลือกชั้นในทันทีที่มี `data` ซ้อนอยู่ แล้ว `actions`
        ที่อยู่ชั้นนอกจะไม่ถูกเห็นเลย

        คำตอบสำหรับ**อ่าน** (GET) ยังเป็น `{ok, data, meta}` ตามเดิม เพราะตารางและ
        ฟอร์มผูกกับ `data` โดยตรง
      properties:
        ok: { type: boolean, const: true }
        message: { type: string, description: ข้อความสรุปผลสำหรับคนอ่าน }
        actions: { $ref: "#/components/schemas/ResponseActions" }
      additionalProperties: true
    ResponseActions:
      type: array
      description: |
        คำสั่งหน้าจอที่เซิร์ฟเวอร์ส่งกลับมา — `ResponseHandler` ของ Now.js อ่านคีย์
        `actions` ที่ระดับบนสุดของคำตอบแล้วทำตามลำดับ

        ใช้กับผลลัพธ์ที่ **ต้องให้ผู้ใช้เห็นเนื้อหาในคำตอบ** ไม่ใช่แค่รู้ว่าสำเร็จ
        เช่นรหัสผ่านที่ระบบสุ่มให้ ซึ่งไม่มีที่อื่นให้ดูย้อนหลังอีกเลย

        ชนิดที่ใช้ได้: `notification` `alert` `modal` `redirect` `update` `render`
        `remove` `class` `attribute` `focus` `scroll` `clipboard` `download` `event`
      items:
        type: object
        required: [type]
        properties:
          type: { type: string }
          message: { type: string }
          title: { type: string }
          content: { type: string }
          level: { type: string, enum: [success, info, warning, error] }
          url: { type: string, description: "`reload` = โหลดใหม่ · ที่เหลือคือปลายทางจริง" }
          target: { type: string, description: "ชื่อ data-table ที่จะโหลดใหม่" }
          action: { type: string, enum: [show, open, close, hide] }
...
    Backup:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string, enum: [site, database, config, full] }
        site_id: { type: [integer, "null"] }
        path: { type: string }
        size_bytes: { type: integer }
        checksum:
          type: string
          description: SHA-256 ของไฟล์ · ไฟล์ที่ไม่มีค่านี้กู้คืนไม่ได้เลยโดยเจตนา
        status: { type: string, enum: [running, ok, failed] }
        status_tone: { type: string, enum: [ok, danger, warn, muted], description: "สีของป้าย status — pill-${status_tone}" }
        created_at: { type: [integer, "null"] }
        site_domain: { type: string }
        destination_id: { type: [integer, "null"] }
        destination_name: { type: string }
        offsite_status: { type: string, enum: [none, ok, failed] }
        offsite_tone: { type: string, enum: [muted, ok, danger], description: "สีของป้าย offsite_status — pill-${offsite_tone}" }
        offsite_at: { type: [integer, "null"] }
        offsite_error: { type: string }
        can_manage:
          type: boolean
          description: สิทธิ์ `backup.manage` ของผู้เรียก
        can_restore:
          type: boolean
          description: สิทธิ์ `backup.restore` ของผู้เรียก
        can_offsite:
          type: boolean
          description: สิทธิ์ `backup.offsite` ของผู้เรียก
    Error:
      type: object
      required: [ok, error]
      properties:
        ok: { type: boolean, const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: รหัสตายตัวที่ฝั่งเรียกใช้ตัดสินใจ — ห้ามเทียบจากข้อความ
              enum:
                - VALIDATION_ERROR
                - UNAUTHENTICATED
                - TWO_FACTOR_REQUIRED
                - PASSWORD_CHANGE_REQUIRED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - CSRF_INVALID
                - RATE_LIMITED
                - QUOTA_EXCEEDED
                - PROTECTED_RESOURCE
                - AGENT_UNAVAILABLE
                - EXECUTION_FAILED
                - BAD_REQUEST
                - METHOD_NOT_ALLOWED
                - INTERNAL_ERROR
            message:
              type: string
              description: ข้อความภาษาไทยที่แสดงให้ผู้ใช้เห็นได้ทันที
            fields:
              type: object
              additionalProperties: { type: string }
              description: ชื่อฟิลด์ => เหตุผลที่ค่านั้นไม่ผ่าน (มีเฉพาะ VALIDATION_ERROR)
  responses:
    Error:
      description: ข้อผิดพลาดตามรูปแบบมาตรฐาน
      headers:
        X-CSRF-Token:
          schema: { type: string }
          description: แนบมากับ 419 เสมอ เพื่อให้ลองใหม่ได้ทันทีโดยไม่ต้องรีเฟรชหน้า
        Retry-After:
          schema: { type: integer }
          description: แนบมากับ 429 เสมอ
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
security:
  - sessionCookie: []